bug: fix memory leak in TabsSessionModel
Before this commit, we only stop observing a session when Tab-switching
happens. It is not enough.

When user switch from Tab-A to Tab-B, we unregister Tab-A, then register
Tab-B. If user close TabTray at the moment, we are still regestering
Tab-B which cause a memory leak.

To fix the issue, to ensure we unregister Tab-B.