tabtray: do not show PrivateMode button in PrivateMode
PrivateMode button is a shortcut to enable PrivateMode from NormalMode
when opening TabTray. We don't need to show the button if we are already
in PrivateMode.

I tried making a opposite button to go back to NormalMode from
PrivateMode. But the UX is not quite good, because user is not easy to
know current mode is Private or Normal - since the TabTray looks same.

And we don't have a Icon for Going to Normal Mode.

Just simply hide the button, until we want to improve the design of
TabTray in PrivateMode.