private: only use legacy session manager
to fix #74

in version 2.7.0, there was a configuration to switch different
SessionManager, base on build type. (`isAcBrowserEngineEnabled()`)

The root cause of ClassCastException was

```
val layoutRes = if (isAcBrowserEngineEnabled()) {
    R.layout.activity_private_mode
} else {
    R.layout.activity_private_mode_legacy
}
```

`R.layout.activity_private_mode_legacy` is using BrowserFragmentLegacy,
and we should not inflate it.

the cast-exception was addressed in commit
912fe4569e671ebc62fac72096786eaa76ceacc5

However, we don't actually use
mozilla.components.browser.session.SessionManager for now. To avoid
making any confusion, this commit remove the new SessionManager for now.