workaround: getDownloadingItems catch any RuntimeException
This is not really fixing the problem. Just ease the pain of exception.

The method `getDownloadingItmes` is for getting information from
DownloadManager and update it to UI. After this commit, except DevBuild,
any RuntimeException will be caught and pretend nothing happend. The
downloading state won't be updated util download finished, namely, the
UI state will not be synced to real download state.

at least, user won't suffer a crash when downloading files.

for issue #57